URL update from /en to root domain
There is a supposition that we may have maxed out number of profiles which is why the root domain results are not working. I am modifying a working URL to see if its profile issue or something wrong with root domain itself. I will revert it either way after I see the result.